pyproject.toml
[tool.poetry]
name = "metac-bot-template"
version = "0.1.0"
description = ""
authors = ["Vasile Popescu <elisescu@elisescu.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
requests = "^2.32.3"
asknews = "^0.9.1"
numpy = "^1.26.0"
openai = "^1.57.4"
python-dotenv = "^1.0.1"
forecasting-tools = "^0.2.23"
psutil = "^5.9.0"
browserbase = "^1.4.0"
playwright = "^1.52.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

